Move exception to public, adjust log level, collapse loops
