+++ modules/compute/src/main/java/org/apache/ignite/internal/compute/IgniteComputeImpl.java
@@ -43,7 +43,7 @@ import org.apache.ignite.compute.IgniteCompute;
 import org.apache.ignite.compute.JobExecution;
 import org.apache.ignite.compute.JobExecutionOptions;
 import org.apache.ignite.compute.JobStatus;
-import org.apache.ignite.internal.compute.exceptions.NodeNotFoundException;
+import org.apache.ignite.compute.NodeNotFoundException;
 import org.apache.ignite.internal.hlc.HybridClock;
 import org.apache.ignite.internal.placementdriver.PlacementDriver;
 import org.apache.ignite.internal.replicator.TablePartitionId;
